Pass filepath to rigger
Previously, when doing a directory import, rigger did not consider that
the importing file would be in the directory being imported, and it
would wind up doing one recursion of self-inclusion.  This change is
part of a small fix that keeps rigger from importing a file into itself
- in the _expandIncludes method, the imported files are checked against
the original file path, and if they match exactly, then the imported
file is marked as not valid for import.